# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Бэкенд «Умного города» — обращения жителей МКД к управляющей компании, плановые работы, справочники домов/УК/аварийных служб. Один инстанс = одна УК.\n\nРоли: representative (представитель УК), dispatcher (диспетчер), resident (житель). Роль определяется самим пользователем на бэкенде — в запросах отдельно не передаётся.",
        "title": "MAXITO API",
        "contact": {
            "name": "MAXITO"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/representative/emergency-services/csv": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обязательные колонки: name, phone. Общий список на всю систему, не по домам. Строки с уже существующим телефоном пропускаются (skipped), не дублируются.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "representative-emergency"
                ],
                "summary": "Массово загрузить аварийные службы из CSV",
                "operationId": "representativeImportEmergencyServicesCSV",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV-файл",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/maxito_internal_services.ImportReport"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/representative/houses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "representative-houses"
                ],
                "summary": "Список всех домов УК",
                "operationId": "representativeListHouses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/maxito_internal_models.House"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "representative-houses"
                ],
                "summary": "Добавить дом",
                "operationId": "representativeCreateHouse",
                "parameters": [
                    {
                        "description": "Адрес, число подъездов; этажность/год/ссылка на чат — необязательно",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_representative.CreateHouseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/maxito_internal_models.House"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "дом с таким адресом и номером уже есть",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/representative/houses/csv": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обязательные колонки: address, entrances_count. Необязательные: number, floors_count, construction_year, chat_invite_link.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "representative-houses"
                ],
                "summary": "Массово добавить дома из CSV",
                "operationId": "representativeImportHousesCSV",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV-файл",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/maxito_internal_services.ImportReport"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/representative/houses/{house_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Любое поле можно не передавать — останется как было.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "representative-houses"
                ],
                "summary": "Редактировать дом",
                "operationId": "representativeUpdateHouse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID дома",
                        "name": "house_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Только те поля, что нужно поменять",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_representative.UpdateHouseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/maxito_internal_models.House"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apidoc.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code — машиночитаемый код ошибки; присутствует только у части\nэндпоинтов (см. описание конкретного эндпоинта), у остальных отсутствует.",
                    "type": "string",
                    "example": "bad_request"
                },
                "error": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
...
        "internal_handlers_representative.CreateHouseRequest": {
            "type": "object",
            "required": [
                "address",
                "entrances_count"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "chat_invite_link": {
                    "description": "необязательно",
                    "type": "string"
                },
                "construction_year": {
                    "description": "необязательно",
                    "type": "integer"
                },
                "entrances_count": {
                    "type": "integer",
                    "minimum": 1
                },
                "floors_count": {
                    "description": "необязательно",
                    "type": "integer"
                },
                "number": {
                    "type": "string"
                }
            }
        },
...
        "internal_handlers_representative.UpdateHouseRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "chat_invite_link": {
                    "type": "string"
                },
                "construction_year": {
                    "type": "integer"
                },
                "entrances_count": {
                    "type": "integer"
                },
                "floors_count": {
                    "type": "integer"
                },
                "number": {
                    "type": "string"
                }
            }
        },
...
        "maxito_internal_models.Appeal": {
            "type": "object",
            "properties": {
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_models.AppealAttachment"
                    }
                },
                "author": {
                    "$ref": "#/definitions/maxito_internal_models.User"
                },
                "author_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "discovered_at": {
                    "description": "когда житель заметил проблему (необязательно)",
                    "type": "string"
                },
                "entrance_number": {
                    "description": "null = весь дом",
                    "type": "integer"
                },
                "house": {
                    "$ref": "#/definitions/maxito_internal_models.House"
                },
                "house_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "importance": {
                    "$ref": "#/definitions/maxito_internal_models.Importance"
                },
                "problem_type": {
                    "$ref": "#/definitions/maxito_internal_models.ProblemType"
                },
                "problem_type_id": {
                    "type": "integer"
                },
                "reason": {
                    "$ref": "#/definitions/maxito_internal_models.Reason"
                },
                "reason_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/maxito_internal_models.AppealStatus"
                },
                "status_changes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_models.AppealStatusChange"
                    }
                },
                "subscriptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_models.AppealSubscription"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "wants_recalculation": {
                    "type": "boolean"
                }
            }
        },
        "maxito_internal_models.AppealAttachment": {
            "type": "object",
            "properties": {
                "appeal_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "maxito_internal_models.AppealStatus": {
            "type": "string",
            "enum": [
                "accepted",
                "in_progress",
                "completed",
                "rejected"
            ],
            "x-enum-varnames": [
                "StatusAccepted",
                "StatusInProgress",
                "StatusCompleted",
                "StatusRejected"
            ]
        },
...
        "maxito_internal_models.AppealSubscription": {
            "type": "object",
            "properties": {
                "appeal_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/maxito_internal_models.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "maxito_internal_models.DispatcherHouse": {
            "type": "object",
            "properties": {
                "assigned_at": {
                    "type": "string"
                },
                "assigned_by": {
                    "description": "Representative who assigned",
                    "type": "integer"
                },
                "assigner": {
                    "$ref": "#/definitions/maxito_internal_models.User"
                },
                "dispatcher": {
                    "$ref": "#/definitions/maxito_internal_models.User"
                },
                "dispatcher_id": {
                    "type": "integer"
                },
                "house": {
                    "$ref": "#/definitions/maxito_internal_models.House"
                },
                "house_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "maxito_internal_models.House": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "appeals": {
                    "description": "Relations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_models.Appeal"
                    }
                },
                "chat_invite_link": {
                    "type": "string"
                },
                "construction_year": {
                    "description": "необязательно",
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "dispatchers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_models.DispatcherHouse"
                    }
                },
                "entrances_count": {
                    "type": "integer"
                },
                "floors_count": {
                    "description": "необязательно",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "notifications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_models.Notification"
                    }
                },
                "number": {
                    "type": "string"
                },
                "residents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_models.Resident"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "maxito_internal_models.Importance": {
            "type": "string",
            "enum": [
                "normal",
                "important"
            ],
            "x-enum-varnames": [
                "ImportanceNormal",
                "ImportanceImportant"
            ]
        },
        "maxito_internal_models.Notification": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/maxito_internal_models.User"
                },
                "author_id": {
                    "type": "integer"
                },
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "ends_at": {
                    "type": "string"
                },
                "entrance_number": {
                    "description": "обязателен при scope_type = entrance",
                    "type": "integer"
                },
                "house": {
                    "$ref": "#/definitions/maxito_internal_models.House"
                },
                "house_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "reason": {
                    "$ref": "#/definitions/maxito_internal_models.Reason"
                },
                "reason_id": {
                    "type": "integer"
                },
                "revoked_at": {
                    "description": "досрочная отмена",
                    "type": "string"
                },
                "scope_type": {
                    "$ref": "#/definitions/maxito_internal_models.NotificationScope"
                },
                "starts_at": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "maxito_internal_models.ProblemType": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "код нужен для простоты определения проблемы",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_critical": {
                    "description": "маркер того, что проблему надо решать срочно и через звонок",
                    "type": "boolean"
                },
                "reasons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_models.Reason"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "maxito_internal_models.Reason": {
            "type": "object",
            "properties": {
                "allows_notification": {
                    "description": "AllowsNotification — можно ли использовать эту причину при создании\nуведомления диспетчером (не при создании обращения жителем — там\nограничений нет). Отсекает абсурдные для планового уведомления\nпричины вроде \"грязный лифт\" — уведомление должно быть про то, что\nможно заранее анонсировать (отключение, неработоспособность), а не\nпро жалобу на текущее состояние.",
                    "type": "boolean"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_other": {
                    "type": "boolean"
                },
                "problem_type_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "maxito_internal_models.Resident": {
            "type": "object",
            "properties": {
                "apartment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "entrance_number": {
                    "type": "integer"
                },
                "house": {
                    "$ref": "#/definitions/maxito_internal_models.House"
                },
                "house_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/maxito_internal_models.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "maxito_internal_models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "dispatcher_houses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_models.DispatcherHouse"
                    }
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "max_user_id": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "resident_profile": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/maxito_internal_models.Resident"
                        }
                    ]
                },
                "role": {
                    "$ref": "#/definitions/maxito_internal_models.Role"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "maxito_internal_services.ImportReport": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "failed": {
                    "type": "integer"
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_services.RowResult"
                    }
                },
                "skipped": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
...
        "maxito_internal_services.RowResult": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "row": {
                    "type": "integer"
                },
                "status": {
                    "description": "created | skipped | error",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT, полученный через POST /auth/max. Передавать как \"Bearer \u0026lt;token\u0026gt;\". Для локальной отладки без MAX бэкенд также принимает X-Max-User-Id (+X-Max-User-Phone при первой привязке), если на сервере включён ALLOW_DEV_HEADERS.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "InternalKey": {
            "description": "Общий секрет между сервером бота и бэкендом — только для /internal/*.",
            "type": "apiKey",
            "name": "X-Internal-Key",
            "in": "header"
        }
    }
}